pyproject.toml
[project]
name = "pxt-cutebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pyserial>=3.5",
]
